Progress commit on moving to Client/Request/Response refactor
